pytest: Test that we can have LSPs modify forward amounts on the fly
This adds a test and two plugins, one for the LSP and one for the user
node, to see if we can allow the LSP to modify forwarded amounts on
the fly. This allows LSPs to leverage a dynamic fee on the fly while
forwarding a payment. We essentially relax the `amt_to_forward` and
`total_msat` verification, to explicitly opt-into this type of
payment. Both the LSP and the client have to agree on the per-HTLC
amounts and the total invoice to be transferred.

In order for both parties to agree there still needs to be a
negotiation phase (how much can the forwarding node drop the value
before the client refuses to accept the payment due to underpayment),
but that phase is omitted here for clarity.

Also notice how the invoice can essentially be modified on the fly as
well, allowing a) dynamic fee negotiations and b) amending the
description to call out the fact that the amount has been modified.

The example here implements a single HTLC, but assuming the LSP tracks
HTLCs (and the amount they already took) correctly it should be simple
to extend this to MPP. Also notice that keysend is a bit difficult for
LSPs unless they charge their fee proportionally on each HTLC
individually, or negotiate with the client to learn the total amount.